Expose Request and RequestMethods to public
