+++ lib/http_mock_adapter.dart
@@ -5,3 +5,4 @@
 library http_mock_adapter;
 
 export 'src/adapters/dio_adapter.dart';
+export 'src/request.dart';
